telemetry: add noop impl. + move constructor in interface files.
